requirements.txt
Django~=3.1.7
djangorestframework~=3.12.2
psycopg2-binary~=2.8.6
django-filter~=2.4.0
pytest~=6.2.2
pytest-django~=4.1.0
model_bakery~=1.2.1
pytz~=2021.1